Refactor provider support
